Extracts DisconnectIntegration
